% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BatchRun}
\alias{BatchRun}
\title{Runs the whole analysis}
\usage{
BatchRun(taxon, ncbi_path = "/usr/local/bin")
}
\arguments{
\item{taxon}{A clade in NCBI taxonomy}

\item{ncbi_path}{Path to where NCBI tools are installed

This starts the analysis; if you already started and the computer crashed, this will pick it up from where it left off. It will run in the current working directory.

The clade in NCBI should map to a single NCBI id; if you have a synonym, try specifying search terms for rentrez::entrez_search so that it finds the correct id.}
}
\value{
Nothing, though see the directory of output
}
\description{
Runs the whole analysis
}
