% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ConcatenateAll}
\alias{ConcatenateAll}
\title{Concatenate the sequences}
\usage{
ConcatenateAll(dna_combined, seqs_final)
}
\arguments{
\item{dna_combined}{sequences}

\item{seqs_final}{Final sequences directory}
}
\value{
TRUE if it worked
}
\description{
Concatenate the sequences
}
