% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CreatePartitionFile}
\alias{CreatePartitionFile}
\title{Create partition file for raxml}
\usage{
CreatePartitionFile(dna_combined, seqs_final)
}
\arguments{
\item{dna_combined}{sequences}

\item{seqs_final}{Final sequences directory}
}
\value{
File name
}
\description{
Create partition file for raxml
}
