% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FixNames}
\alias{FixNames}
\title{Corrects taxon names}
\usage{
FixNames(reduced, cid)
}
\arguments{
\item{reduced}{Phylota}

\item{cid}{Cluster id}
}
\value{
fixed names
}
\description{
Corrects taxon names
}
