% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KeepClusters}
\alias{KeepClusters}
\title{Find clusters to keep}
\usage{
KeepClusters(cids, n_taxa, mintaxa)
}
\arguments{
\item{cids}{cluster ids}

\item{n_taxa}{number of taxa per cluster}

\item{mintaxa}{minimum number of taxa per cluster}
}
\value{
vector of clusters to use
}
\description{
Find clusters to keep
}
