% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ProcessSequencesByGeneConcat}
\alias{ProcessSequencesByGeneConcat}
\title{Process the output of phylotaR}
\usage{
ProcessSequencesByGeneConcat(inputs, seqs_processed, seqs_raw)
}
\arguments{
\item{inputs}{input info}

\item{seqs_processed}{Output directory}

\item{seqs_raw}{Raw sequences directory}
}
\value{
output files
}
\description{
Process the output of phylotaR
}
