% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ProcessSequencesByGeneSingle}
\alias{ProcessSequencesByGeneSingle}
\title{Process the output of phylotaR by gene}
\usage{
ProcessSequencesByGeneSingle(
  inputs,
  seqs_raw,
  seqs_processed,
  seqs_direct_download
)
}
\arguments{
\item{inputs}{input info}

\item{seqs_raw}{Raw sequences directory}

\item{seqs_processed}{Output directory}

\item{seqs_direct_download}{Direct download sequences directory}
}
\value{
output files
}
\description{
Process the output of phylotaR by gene
}
