% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RemoveGappy}
\alias{RemoveGappy}
\title{Remove gappy columns}
\usage{
RemoveGappy(inputs, seqs_gappy_removed, seqs_processed)
}
\arguments{
\item{inputs}{input info}

\item{seqs_gappy_removed}{Output directory}

\item{seqs_processed}{Processed sequences directory}
}
\value{
output files
}
\description{
Remove gappy columns
}
