% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RunPhylotaR}
\alias{RunPhylotaR}
\title{Runs phylotaR}
\usage{
RunPhylotaR(wd, txid, ncbi_dr)
}
\arguments{
\item{wd}{Working directory}

\item{txid}{NCBI taxon id}

\item{ncbi_dr}{Path to where NCBI tools are installed}
}
\value{
Phylota
}
\description{
Runs phylotaR
}
