% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CreatePartitionFile}
\alias{CreatePartitionFile}
\title{Create partition file for raxml}
\usage{
CreatePartitionFile(dna_combined, seqs_final)
}
\arguments{
\item{dna_combined}{sequences}

\item{seqs_final}{Final sequences directory}
}
\value{
File name
}
\description{
This makes two files: one with partitions for cpDNA, mtDNA, 18S, 28S, and the rest (partitions.txt) and one with each gene in its own partition (partitions_bygene.txt)
}
