% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RunRaxml}
\alias{RunRaxml}
\title{Run raxml}
\usage{
RunRaxml(seqs_final, raxml, ...)
}
\arguments{
\item{seqs_final}{Final sequences directory}

\item{raxml}{RAxML name and perhaps path}

\item{...}{Other arguments to make targets wait before runnnig}
}
\value{
TRUE if it worked
}
\description{
Run raxml
}
