% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SaveGenes}
\alias{SaveGenes}
\title{Save genes to a file}
\usage{
SaveGenes(reduced, seqs_raw)
}
\arguments{
\item{reduced}{Clusters}

\item{seqs_raw}{Raw sequences directory}
}
\value{
Output files
}
\description{
Save genes to a file
}
